backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP clients
httpx==0.25.2
requests==2.31.0

# AI/LLM
google-generativeai==0.3.1

# Environment & Config
python-dotenv==1.0.0

# Retry logic
tenacity==8.2.3

# Date handling
python-dateutil==2.8.2

# CORS
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
